chore(gather-packages-test): refactors workspaces & package finding into own file
This code is ostensibly complex and elides information that we need for
doctor.
